Add local cloudbuild support, modify cloudbuild script to accept namespace as arg
