--- runtimes-common-testing/src/main/java/com/google/cloud/runtimes/jetty/runtimes_common_test/DefaultServlet.java
@@ -1,21 +1,0 @@
-package com.google.cloud.runtimes.jetty.runtimes_common_test;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-@WebServlet(urlPatterns = {"/*"})
-public class DefaultServlet extends HttpServlet {
-
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-      throws ServletException, IOException {
-    resp.setContentType("text/plain");
-    resp.getWriter().println("Hello World!");
-  }
-
-}
